requirements.txt
APScheduler==3.11.0
aiohttp==3.11.11
pydantic==2.10.4
tzlocal==5.2
discord.py==2.4.0
pytz==2024.1